change controller for error
