+++ src/main/java/com/fpt/t1708e/photoplatform/controller/studio/AlbumController.java
@@ -126,34 +126,4 @@ public class AlbumController {
         }
         return "redirect:album/list";
     }
-
-    @RequestMapping(method = RequestMethod.GET, value = "/error/404")
-    public String error1(Model model) throws RemoteException {
-        return "manager/admin/error/404";
-    }
-
-    @RequestMapping(method = RequestMethod.GET, value = "/error/500")
-    public String error2(Model model) throws RemoteException {
-        return "manager/admin/error/500";
-    }
-
-    @RequestMapping(method = RequestMethod.GET, value = "/error/other")
-    public String error3(Model model) throws RemoteException {
-        return "manager/admin/error/other";
-    }
-
-    @RequestMapping(method = RequestMethod.GET, value = "/error/customer/404")
-    public String errorCustomer1(Model model) throws RemoteException {
-        return "manager/error/404";
-    }
-
-    @RequestMapping(method = RequestMethod.GET, value = "/error/customer/500")
-    public String errorCustomer2(Model model) throws RemoteException {
-        return "manager/error/500";
-    }
-
-    @RequestMapping(method = RequestMethod.GET, value = "/error/customer/other")
-    public String errorCustomer3(Model model) throws RemoteException {
-        return "manager/error/other";
-    }
 }
